// gaz/app/src/app/models/pos.model.ts
export interface PosModel {
  code: string;
  branch_office: string;
  name: string;
  establishment: string;
  phone: string;
  address: string;
  neighborhood: string;
  city_code: string;
  city_name: string;
  department_code: string;
  department_name: string;
  id_supervisor: string;
  id_price_list: string;
  id_discount_list: string;
}